NewAttempt/requirements.txt
python-chess>=1.9.0
numpy>=1.24.0
torch>=2.0.0
modal>=0.56.0
tqdm>=4.65.0
